Implement Tags Button and Methods on Context
